avoid dependency from openapi-manager to libnvme
Created using spr 1.3.6-beta.1
