+++ sled-agent/api/Cargo.toml
@@ -18,5 +18,4 @@ schemars.workspace = true
 serde.workspace = true
 sled-agent-types.workspace = true
 sled-hardware-types.workspace = true
-sled-storage.workspace = true
 uuid.workspace = true
