+++ sled-agent/src/sim/storage.rs
@@ -19,15 +19,15 @@ use dropshot::HandlerTaskMode;
 use dropshot::HttpError;
 use futures::lock::Mutex;
 use omicron_common::disk::DiskIdentity;
+use omicron_common::disk::DiskManagementStatus;
 use omicron_common::disk::DiskVariant;
+use omicron_common::disk::DisksManagementResult;
 use omicron_common::disk::OmicronPhysicalDisksConfig;
 use omicron_uuid_kinds::GenericUuid;
 use omicron_uuid_kinds::InstanceUuid;
 use omicron_uuid_kinds::OmicronZoneUuid;
 use omicron_uuid_kinds::ZpoolUuid;
 use propolis_client::types::VolumeConstructionRequest;
-use sled_storage::resources::DiskManagementStatus;
-use sled_storage::resources::DisksManagementResult;
 use slog::Logger;
 use std::collections::HashMap;
 use std::collections::HashSet;
