+++ sled-agent/types/Cargo.toml
@@ -28,7 +28,6 @@ serde_human_bytes.workspace = true
 serde_json.workspace = true
 sha3.workspace = true
 sled-hardware-types.workspace = true
-sled-storage.workspace = true
 slog.workspace = true
 thiserror.workspace = true
 toml.workspace = true
